Minor improvement to Straightness (all versions thereof)
When adding up segment lengths, iterate over edge particles by
following the actual bonds from one particle to the next, rather
than just by theta. That's more reliable and accurate because
occasionally the edge gets kinked and doubles back on itself,
and then the paths don't go in order of theta, so it would not
give the true path length.