Bond remodeling: enable special constraint lambda configuration
Normally, there's one one lambda (k) for the bond-angle
constraint when an edge bond is being remodeled (i.e. when
both particles are edge particles), and a second lambda for
all the other cases.

But now, enable a special experiment where k_edge_bond_angle
applies not only when an edge bond is being remodeled, but
when ANY bond involving an edge particle is being remodeled,
i.e. not only the bonds representing the leading edge itself,
but also the bonds connecting the edge particles to the next
inner tier of particles.