Chemical species - proof of concept
Fill 10 cells with chemical B, at a value of 1.0 (for convenience
because it fits the color scale, which goes from 0 to 1).

As it diffuses to other cells, keep topping off the original
cells to 1.0, otherwise the chemical spreads too thin and
disappears. This way, the level builds up.

Color the cells according to the level of the chemical, and
re-color them every single timestep.