Yolk-evl spring constant: settle on 5
Don't know what I was thinking. I tested this again and could
not see any jitter in the cell-division sims when increasing
this value to 5. So go ahead and do that, to bring +/- cell
division sims into alignment, so all their parameters are
comparable. (And therefore no need to mess with dt.)