Plotting: minor refactor; generate _phi array one level up for visibility
It was already *stored* globally for export; but really just
belonged to the one plot, "Leading edge phi", plotting phi
on the y-axis vs. time; and was generated inside the function
for that plot. But now extract that from that particular
plotting function and do it before calling all the various plots,
just so its generation is independent of the order of execution of
the plots, and it's plainly visible at that level of the code.
Just like the _timesteps array.

Now we'll use it in other plots (as the x-axis).