Plotting: New plots: tension over a much finer timescale
In contrast to my older tension plots (binned local tension vs.
phi *and* time), this just a single scalar measurement for
median tension over all particles, and a second single scalar
measurement for median tension over just the leading edge
particles, both plotted against time; this allows a simpler
plot that can capture many more time points. (I.e., sacrificing
spatial granularity in favor of temporal granularity.)

Plot two ways: vs timestep, like all the other plots, and in a
second plot vs. leading edge mean phi, i.e. epiboly progress as a
spatial representation of time, just like % epiboly but in
terms of angle (nice and linear) instead of in terms of % of
embryo axis like biologists do it (not linear). The two are
the same plot but the x-axes are distorted relative to one
another. I believe plotting vs. epiboly progress rather than
raw time is preferable.

Try this out for awhile, and if it's good, maybe convert all
the other similar plots of value-vs-time, to this same method.

Note, looking at my older tension plots, I notice they only
measure internal EVL cells and not the leading edge! Can't
remember if that was because I was just too quick-and-dirty
and never came back to refine it; or if it was because of
some difficulty in including the leading edge cells. If I
decide to publish those, better go back and take a look at
that and clean it up!