Data import: refactor (anticipating post-process plotting)
Will want to plot data from multiple runs together on one
set of axes. To do that, will need to load in multiple sets
of data from disparate directories. Refactor out this code
(previously used only when re-initializing an in-progress
simulation from incomplete exported data, such as after a
crash) so that it can be accessed separately.