Plotting: refactor data plotting away from data generation
Another step toward post-process plotting. When doing it
post-process, the data will already exist and will just need
to be passed into the plotting function.

For now, just applied to Straightness Index and Lopsidedness
plots, but can easily deploy more widely as needed. (Probably
should do this at least for all the plots that are v. time, since
they're all so similar and this would vastly reduce redundancy
and simplify my code. But hold that thought for now...)