Plotting: Enable plotting multiple datasets on one Figure/Axes
Pass a list of datasets instead of just a single one. Plot all
of them on the same graph. Also provide for different line
style options (all the same, or all different); eventually I'll
also need a custom option, where the caller specifies the style
for each dataset explicitly.

For now, still only passing a single dataset: it's good enough
that it doesn't break anything. To test that it works as
intended, I'll need to arrange to actually load in multiple
datasets and pass them.