Plotting multiple: refactor tensions; increase calling flexibility
Refactor one more plot into data generation vs. plotting: the
tensions_v_microtime plot.

To do this, needed to actually use the multiple-datasets-per-Axes
feature, not in the context of post-processing, but simply in
the normal real-time simulation, because this plot has two
datasets plotted on it. Good test case!

And to do this, needed more flexibility: implement the option
for explicit, per-dataset line formatting as anticipated (sooner
than anticipated!), and also per-dataset labels for the legend.
As well as the ability to specify the presence and position of
an axvline.