Plotting multiple: Flex x axis time|phi; and refactor progress plot
Allow caller to specify whether x axis will be vs. time, or vs. phi,
rather than always looking at the balanced_force_control config.
This is more flexible, lets us override that. E.g., for epiboly
progress plot, which is always vs. time. And anticipates that in
post-progress plotting, we'll need to pass the x data for each
set of y data anyway, so no need to restrict it to only phi.

Refactor the epiboly progress plot (phi vs. time) to do it this way.