Plotting multiple: make composite plots generic
In the generic sense of "generic", not the computer
science sense of "generic"!

Pass the multiple-simulation data to the function that
bins it and takes the median of each bin, without
that function needing to know what the data represents.
So that it can be reused for various plots.

You just pass in the data, the treatment values, and
a couple of data-dependent specifications for the plot,
and show_composite_medians() should take care of the
rest.