Plotting multiple: mark a dataset to exclude from composite
If simulation set specified as all simulations within a
given enclosing directory, enable excluding any
simulation whose name starts with "skip". That way,
I can keep that data together in the same folder with
the other runs, but not include it in the composite
plot. For example, if a particular run results in some
kind of corrupted result that would mess up the data.