Plotting multiple: control plot colors for bool groupings
When grouping into two sets according to a bool config
var, select which one should appear first in the
legend and use cycler color C0. Sometimes the default
(False comes first because it's numerically 0) is
counterintuitive and feels awkward in the context of
other accompanying plots. Just depending on the
semantics of the bool var. This lets us flip them.