Config: let k_edge_bond_angle default to 4, not 3.75
I have been running it this way for quite some time
now; might as well make it official. When plotting
a family of curves for runs at different values of
this parameter, it was just more convenient and
sensible to have "4" in our plot instead of the
overly precise "3.75".

This will now be the default "normal" value to use
when we're varying other parameters.