Config var rename: change constraint k's to lambdas
Finally, rectifying bad decision that originally named
these "k" because "lambda" is a python reserved word.
(But only when all by itself.)

Now I have so many k parameters that it's just too
confusing; plus, for the strength of my constraints,
I'm calling it lambda in the manuscript, so do the
same here. Reserve "k" for spring constants (which
are called "k" by Tissue Forge as well) and one other
minor thing (k_particle_diffusion, which I'm not
using anymore anyway, and isn't causing any
confusion).

Unfortunately, I cannot change it where the variable
names are used as dictionary keys in the metadata
file, because it would invalidate all the existing
data files when I try to do plots of multiple runs
together. I.e. for backward compatibility. At least
not without getting distracted by having to fix the
code to support both, and interchangeably, which is
not a good use of time! Have to just live with this
for now, until I'm SURE I won't be reanalyzing those
same files anymore. (Like post-publication!)