Video: special camera control for lineage tracing
When tracing cells in a given tier, suppress camera
rotation if tier > 1 so that the labeled cells don't
rotate out of view.

When tracing cells in a patch, rotate the camera as
usual, but not all the way to the vegetal pole; stop
half-way there. Furthermore, only bother to make a
movie for the front side (the others definitely
are not interesting, so save myself the trouble of
discarding them manually); and, when deleting the
screenshots after the movie is finished, retain the
first and last frame of the front side (these
definitely ARE interesting, so save myself the trouble
of capturing them manually).