Plotting multiple: Show division cessation in plots +/- cell division
In consensus plots, typically all simulations represented
either have cell division, or don't. The vertical line
was added only for simulations with cell division.

In the exception, those consensus plots comparing
simulations with and without cell division, the behavior
was to simply not draw such a vertical line.

Now instead, add the line even though it applies only
to some of the simulations in the plot (the ones with
cell division). Also include a function argument to
make this optional, so I can show it plots where I
want to draw attention to it, and leave it out of
plots where it's an irrelevant distraction.