Balanced force equilibration - tweak the plotting
I was preventing plots v phi. But actually it's helpful
to see, just need to keep in mind that the part of the
plot during the equilibration phase is messed up. Just
make sure the time-based x-axis types are also included.
But this is only for my own use anyway. If I ever decide
to add this to the model and use it across the board,
I wouldn't be including the equilibration phase in
the plot.