Fix cryptic type-hinting syntax error
Yikes, I was type-hinting dicts with dict[key: value]
instead of dict[key, value] in several places, and the
type-checker never alerted me! It's a wonder this worked
at all, because it's completely wrong, but they all
worked because type-hinting is irrelevant at runtime
anyway; and the type-checker missed them because I
wasn't then indexing into the dict with literal keys,
but with variable keys – something the type-checker
doesn't seem to really get.

(Although that's also weird because in other spots,
the type-checker complained about my variable keys,
but not in these places.)

It finally complained while I'm developing something new,
where I'm indexing into the badly-type-hinted dicts
using literal keys, and it could detect the mismatch.

(Edit: It turned out that's not why it complained;
correcting the syntax in the new code didn't fix the
complaints in the code I'm working on; the complaint
seems to be a PyCharm bug, having nothing to do with
the syntax error! I had to just suppress it. It was
only a coincidence that this led me to discover the
syntax error. Which apparently had no negative effects
at all. Well, it's fixed now.)

Of the old ones, tf.utils.py has just old unneeded
cruft so removed it entirely; fixed the other items
in plotting.py.