Plotting post-process: Consensus medians: account for truncated data (Model 1)
Model 1 sims end over a range of different
leading-edge-mean-phi. Need to first account for those
"missing" x values before we can interpolate this data.
Can't just pass a ragged array to np.median(), because
numpy can't do ragged arrays. Need to implement a way
to generated interpolated data sets over different
domains and then at each x, take the medians of just
the ones that have survived to there.
For now, use this only for Model 1.

Also, optionally, account for bias arising out of the
truncation. As some sims drop off, the ones that
continue further may be unrepresentative, i.e.
survivorship bias. Have to use this on a case-by-case
basis, depending on the particular metric. Implemented
here, remove low values, on the assumption that the
survivors have a low bias. This is the case for
the lopsidedness metric. For straightness index, I'd
want to probably do the opposite. For now, I probably
end up not using consensus plots for either of those
metrics anyway. The other metrics don't appear to
need such a correction.