Tweak yolk-evl spring constant
I've been running it with this slightly modified value
for many months, but just never checked it in. With
certain parameter values the leading edge was digging
itself in and the bonds and even cells would disappear
into the yolk cell a little bit. Backing off on this
value remedied it, and hasn't had any adverse effects
anywhere, so let this be the new normal.