Plotting: Add ability to specify final plot size
This is for post-processed plots, for
better publication-quality.

Specify the desired final height of the
whole Figure in inches; it will be scaled
so that the axis labels and tick marks
aren't shrunk; the actual plot takes up the
remaining space. This makes a much more
readable plot.

With this, slightly change when to plot
the legend outside the plot region.
Instead of using the value of legend_loc
as a flag as before, now this function is
nested inside the new adjust_figsize(). So
only move the legend when doing a size
adjustment, but then just always do it,
ignoring legend_loc.