Plotting multiple: Add a new use-case
Previously did not extract, for each
simulation output file, whether it was
Model 1 or Model 2 — UNLESS I was
intentionally combining both in a single
plot in order to contrast them against one
another. But it was safe to assume in all
other cases that the data was from Model 2.
(Because I just didn't have any use cases
where I needed to plot Model 1 alone.)
Thus it was not possible to plot Model 1
correctly UNLESS Model 2 was also present
(which it always was).

But now I've done a new experiment and
needed a plot with just Model 1. So, added
the ability to extract the model number
from the simulation data. Now the plot
works. Also made the query ("which model
is this simulation?") simpler to read and
understand.