Plotting multiple: Add another new use case
Allow for plotting truncated domains even
with Model 2. It works the same, but
previously I reserved it for Model 1, and
continued using the older simpler version
with Model 2, because Model 2 didn't need
this feature, so I continued with the older
version to be conservative.

But now I added a new experiment and plot
to the manuscript, in which I look at just
Model 2, but some of the treatments don't
complete epiboly and therefore have to be
terminated manually, resulting in a
truncated phi domain. Same exact thing as
with Model 1. So use the same approach to
plot these. It actually simplifies the
code since all plots now treated the same.

Old code retained (commented out) since
it represents an easier-to-understand
version.