Plotting: improve axvline and axhline style
For those marker lines in plots, set
linewidth to 1.0 instead of 0.5. (Journal
requests this as smallest final linewidth
after scaling, and since I'm now plotting
at the final desired size no scaling should
be needed, so this should be right.)

The dotted didn't look great, and the dots
might be too small, so switch to dashed.

And the 1.0 point dashes are jarringly
bold, so back off the color from black
to 0.25 gray. That looks decent.